requirements.txt
streamlit
redis
beautifulsoup4
requests
langchain
python-dotenv
langgraph